+++ src/dates.rs
@@ -15,6 +15,25 @@ pub fn date_to_string(date: NaiveDate, separator: &str, leading_zeros: bool) -> 
         + &day.to_string()
 }
 
+//TODO: Documention
+pub fn date_month_to_string(date: NaiveDate) -> &'static str {
+    match date.month0() {
+        0 => "January",
+        1 => "February",
+        2 => "March",
+        3 => "April",
+        4 => "May",
+        5 => "June",
+        6 => "July",
+        7 => "August",
+        8 => "September",
+        9 => "October",
+        10 => "November",
+        11 => "December",
+        _ => unreachable!(),
+    }
+}
+
 pub fn first() -> NaiveDate {
     NaiveDate::from_ymd_opt(1978, 6, 19)
         .expect("Static date failed to parse. This error should never occur.")
