+++ src/proxy.rs
@@ -1,42 +1,14 @@
-use chrono::{Datelike, NaiveDate};
+use chrono::NaiveDate;
 use reqwest::Client;
 
-use crate::dates::date_to_string;
+use crate::SourceAPI;
 
-pub fn webpage_proxied(date: NaiveDate, proxy: Option<&str>) -> String {
-    let url = webpage_unproxied(date);
+//TODO Inline this???
+pub fn webpage_proxied(date: NaiveDate, proxy: Option<&str>, api: SourceAPI) -> String {
+    let url = api.get_page_url(date);
     match proxy {
         None => url,
         Some(proxy) => proxy.to_string() + "?" + &url,
-    }
-}
-
-fn webpage_unproxied(date: NaiveDate) -> String {
-    let month_string = date_month_to_string(date);
-    let date_string = date_to_string(date, "-", false);
-    format!(
-        "https://garfield.fandom.com/wiki/Garfield,_{month}_{year}_comic_strips?file={date}.gif",
-        month = month_string,
-        year = date.year(),
-        date = date_string,
-    )
-}
-
-fn date_month_to_string(date: NaiveDate) -> &'static str {
-    match date.month0() {
-        0 => "January",
-        1 => "February",
-        2 => "March",
-        3 => "April",
-        4 => "May",
-        5 => "June",
-        6 => "July",
-        7 => "August",
-        8 => "September",
-        9 => "October",
-        10 => "November",
-        11 => "December",
-        _ => unreachable!(),
     }
 }
